// M1-CHAN/M1-PARTIAL/M1-CHAN/ClientApp/src/app/interfaces/ItemInterface.ts
export interface ItemInterface {
  Id?: number | null;
  ItemId?: number | null;
  ItemCode?: string | null;
  ItemName?: string | null;
  ItemBrand?: string | null;
  UnitPrice?: number | null;
  ItemSize?: string | null;
  ItemType?: string | null;
  ItemVariant?: string | null;
  ItemRemarks?: string | null;
  CustomerId?: number | null;
}